backend/requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.0
pydantic>=2.3.0
pydantic-settings>=2.0.3
python-dotenv>=1.0.0
loguru>=0.7.0
python-multipart>=0.0.6
email-validator>=2.0.0
httpx>=0.25.0
passlib>=1.7.4
python-jose>=3.3.0
bcrypt>=4.0.1
starlette>=0.27.0
pandas>=2.1.0
openpyxl>=3.1.2
xlrd>=2.0.1
pytest>=7.4.2
pytest-asyncio>=0.21.1 